requirements.txt
Flask==3.0.3
supabase==2.27.1
python-dotenv==1.2.1
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
face-recognition>=1.3.0
pgvector>=0.2.0
requests>=2.31.0
werkzeug>=3.0.0
PyYAML>=6.0
# Note: inference-sdk may need to be installed separately if available
# pip install inference-sdk (if package is available from Roboflow)

